07-31-2020: Status update on progress w/extra attachment...
Along with other tidying up done to this file, a new section was added. Line "2" was expanded to include some problems outside of my hands. I hope that a fix will be made soon but I don't want to be a bother. If you have any information on this bug before I decide to submit a bug report/issue please let me know.