First Version Of Fork Request! - "PullRequest1aVersionUncleaned"
I am so glad to finally being able to submit my work to the real repository and see if it gets accepted. This is the version that will be submitted but hasn't had the unuseful comments removed. This is just being committed before the rest to catch any issues that even with my vigorous testing did not catch.