Second Version Of Pull Request! - "PullRequest1aCleaned"
I am so glad to finally being able to submit my work to the real repository and see if it gets accepted. I have spent a lot of time on this and can't wait to share it. This is the version that will be submitted and it has the unuseful comments removed.